requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
schedule==1.2.0
requests>=2.25.1
selenium==4.15.2
webdriver-manager==4.0.1
python-dateutil>=2.8.2
beautifulsoup4>=4.9.3
feedparser>=6.0.2
torch==2.0.0  # For Meta Llama
transformers==4.30.0  # For Meta Llama
